+++ config/configtelemetry/configtelemetry_test.go
@@ -67,54 +67,6 @@ func TestParseFrom(t *testing.T) {
 	}
 }
 
-func TestLevelSet(t *testing.T) {
-	tests := []struct {
-		str   string
-		level Level
-		err   bool
-	}{
-		{
-			str:   "",
-			level: LevelNone,
-			err:   true,
-		},
-		{
-			str:   "other_string",
-			level: LevelNone,
-			err:   true,
-		},
-		{
-			str:   levelNoneStr,
-			level: LevelNone,
-		},
-		{
-			str:   levelBasicStr,
-			level: LevelBasic,
-		},
-		{
-			str:   levelNormalStr,
-			level: LevelNormal,
-		},
-		{
-			str:   levelDetailedStr,
-			level: LevelDetailed,
-		},
-	}
-	for _, test := range tests {
-		t.Run(test.str, func(t *testing.T) {
-			lvl := new(Level)
-			err := lvl.Set(test.str)
-			if test.err {
-				assert.Error(t, err)
-				assert.Equal(t, LevelBasic, *lvl)
-			} else {
-				assert.NoError(t, err)
-				assert.Equal(t, test.level, *lvl)
-			}
-		})
-	}
-}
-
 func TestLevelString(t *testing.T) {
 	tests := []struct {
 		str   string