+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## 🛑 Breaking changes 🛑
+
+- Deprecate `service/defaultcomponents` go package (#4622)
+
 ## 💡 Enhancements 💡
 
 - Transform configmapprovider.Retrieved interface to a struct (#4789)
@@ -21,6 +25,7 @@
 
 - Invalid requests now return an appropriate unsupported (`405`) or method not allowed (`415`) response (#4735)
 - `client.Info`: Add Host property for Metadata (#4736)
+- Add support for reloading TLS certificates (#4737)
 
 ## v0.43.1 Beta